secret_message.py
Several images which have names in an alphanumeric pattern are stored in a folder. The aim of this mini project is to rename the files in such a way that the numeric part of the names are removed and the files are then reordered. The image files have a hidden message which is revealed on reordering after the files are renamed.